cloud/cloud-functions/stocks_history/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery==1.8.0